Testing IMU
The IMU was not spitting out the data, probably because of the SPI init
function (I copied the one from the example and it works).
We also using the filters that used to be for the SLUGS CUBE before,
however, we should switch to the new scales and biases.
We still need to figure out if we should use high sample rate and not
just 100hz.