First Flight In full configuration
#Before first flight
We found that the aileron servos throw were reversed on the Phonix. One
option is to reverse them on the RC but then it will be
inverted for the auto-pilot and needed to reprogramed the AUAV3 for the
second configuration. We decided to turn them upside down
such that it will be as it is on the Bixler2 and we will not have two
configurations.
#During first flight
The last one will be the one to analyze.
The Airplane model started the take off by a person who threw it while
the RC Pilot moved the throttle to half of the travel.
In manual mode, the RC Pilot was in full control and it seems that the
simple logic of passthrough with all the internal conversions
worked just fine.
The weather was warm with a wind of less than 7m/sec. However, we flew
around and it seemed like we had a turbulent above the east part of the
second hill.
And in a specific spot, we could observe that the airplane flew through
air-pocket.
After a couple of turns around the take-off location, we decided to
switch to stabilized mode. Switching to stabilized mode caused the
airplane to drop the right wing and start
diving. We switch back and forth between manual mode and stabilized mode
and we observed the same outcome.
The last try was right above the air-pocket spot and the airplane try to
bank to the right and dive but it seems that the
sheer wind was strong enough to break the right wing. The airplane
continued to climb and then gain a steep dive to the ground with no
serious damage to it.
#Before second flight
The second drone (Phoenix) was having issues with the GPS receiver and
we replace it with the one on the Bixler2 and since we saw that the 3dr
transceiver used for the
telemetry was much more stable before. We then switch between those as
well.
#During second flight
We followed the same procedure on the second flight but now we tried to
be more careful and avoid switching modes above that specific unstable
spot.
Surprisingly when we switch to Stabilized mode it flew stable though
turning and dropping altitude a little bit. When we switch to Autonomous
mode it was stable and hovered around us when it actually supposes to
follow a mission plan with two waypoints.
We switch back and forth between three modes and the same behavior
occurred. The Phoenix landed safely.
We need to analyze the Log files.